refactor: move dto
