+++ apps/person/src/entities/index.ts
@@ -1,4 +1,4 @@
-export * from './person.entity';
-export * from './film_role.entity';
-export * from './person_film_role.entity';
-export * from './film_person.entity';
+export * from './person/person.entity';
+export * from './film-role/film_role.entity';
+export * from './person/person_film_role.entity';
+export * from './person/film_person.entity';
